+++ server/src/main/java/org/opensearch/snapshots/SnapshotsService.java
@@ -134,6 +134,7 @@ import java.util.stream.Stream;
 import static java.util.Collections.emptySet;
 import static java.util.Collections.unmodifiableList;
 import static org.opensearch.cluster.SnapshotsInProgress.completed;
+import static org.opensearch.repositories.blobstore.BlobStoreRepository.REMOTE_STORE_INDEX_SHALLOW_COPY;
 import static org.opensearch.snapshots.SnapshotUtils.validateSnapshotsBackingAnyIndex;
 
 /**
@@ -337,6 +338,7 @@ public class SnapshotsService extends AbstractLifecycleComponent implements Clus
                 );
 
                 logger.trace("[{}][{}] creating snapshot for indices [{}]", repositoryName, snapshotName, indices);
+                boolean remoteStoreIndexShallowCopy = REMOTE_STORE_INDEX_SHALLOW_COPY.get(repository.getMetadata().settings());
                 newEntry = new SnapshotsInProgress.Entry(
                     new Snapshot(repositoryName, snapshotId),
                     request.includeGlobalState(),
@@ -348,7 +350,8 @@ public class SnapshotsService extends AbstractLifecycleComponent implements Clus
                     RepositoryData.UNKNOWN_REPO_GEN,
                     ImmutableOpenMap.of(),
                     userMeta,
-                    Version.CURRENT
+                    Version.CURRENT,
+                    remoteStoreIndexShallowCopy
                 );
                 initializingSnapshots.add(newEntry.snapshot());
                 snapshots = SnapshotsInProgress.of(Collections.singletonList(newEntry));
@@ -425,6 +428,7 @@ public class SnapshotsService extends AbstractLifecycleComponent implements Clus
         // retries
         final SnapshotId snapshotId = new SnapshotId(snapshotName, UUIDs.randomBase64UUID()); // new UUID for the snapshot
         Repository repository = repositoriesService.repository(request.repository());
+
         if (repository.isReadOnly()) {
             listener.onFailure(new RepositoryException(repository.getMetadata().name(), "cannot create snapshot in a readonly repository"));
             return;
@@ -514,6 +518,8 @@ public class SnapshotsService extends AbstractLifecycleComponent implements Clus
                         );
                     }
                 }
+
+                boolean remoteStoreIndexShallowCopy = REMOTE_STORE_INDEX_SHALLOW_COPY.get(repository.getMetadata().settings());
                 newEntry = SnapshotsInProgress.startedEntry(
                     new Snapshot(repositoryName, snapshotId),
                     request.includeGlobalState(),
@@ -524,7 +530,8 @@ public class SnapshotsService extends AbstractLifecycleComponent implements Clus
                     repositoryData.getGenId(),
                     shards,
                     userMeta,
-                    version
+                    version,
+                    remoteStoreIndexShallowCopy
                 );
                 final List<SnapshotsInProgress.Entry> newEntries = new ArrayList<>(runningSnapshots);
                 newEntries.add(newEntry);
@@ -1034,7 +1041,8 @@ public class SnapshotsService extends AbstractLifecycleComponent implements Clus
                                 repositoryData.getGenId(),
                                 ImmutableOpenMap.of(),
                                 snapshot.userMetadata(),
-                                version
+                                version,
+                                snapshot.remoteStoreIndexShallowCopy()
                             ),
                             clusterState.metadata(),
                             repositoryData
@@ -1840,7 +1848,8 @@ public class SnapshotsService extends AbstractLifecycleComponent implements Clus
                 entry.partial() ? shardGenerations.totalShards() : entry.shards().size(),
                 shardFailures,
                 entry.includeGlobalState(),
-                entry.userMetadata()
+                entry.userMetadata(),
+                entry.remoteStoreIndexShallowCopy()
             );
             final StepListener<Metadata> metadataListener = new StepListener<>();
             final Repository repo = repositoriesService.repository(snapshot.getRepository());